toolbox: prevent mounted snapshots from being gc'ed
The toolbox uses ctr under the hood if available. The 'ctr image mount'
command by default adds a 1 day lease on the created snapshot that is
mounted as the upperdir. Therfore the snapshot will be gc'ed and the
toolbox not function correctly anymore as the upperdir does not longer
exist.